assistant_project/requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
transformers>=4.35.0
torch>=2.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
scipy>=1.10.0
numpy>=1.24.0
requests>=2.31.0
fasttext-wheel>=0.9.2
sentencepiece>=0.1.99
sacremoses>=0.1.1
